+++ contracts/mocks/compound/CTokenInterfaces.sol
@@ -33,7 +33,7 @@ abstract contract CErc20Interface is CErc20Storage {
     /**
      * This is used to deposit into Compound with CErc20 tokens
      */
-    function mint(uint mintAmount) external virtual returns (uint);
+    function mint(uint256 mintAmount) external virtual returns (uint);
 
-    function redeem(uint redeemTokens) external virtual returns (uint);
+    function redeem(uint256 redeemTokens) external virtual returns (uint);
 }
